+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 See README.md before updating this file.
 
 ## Unreleased [#](https://github.com/enova/landable/compare/v1.7.1.rc1...master)
+* Screenshots by revision [#7]
 
 ## 1.7.1.rc1 [#](https://github.com/enova/landable/compare/v1.7.0...v1.7.1.rc1)
 * Allow for revisions for Templates [#3]