+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 See README.md before updating this file.
 
 ## Unreleased [#](https://github.com/enova/landable/compare/v1.7.1.rc1...master)
+* Some fixes to the DB tests [#9]
 * Rails 4.x support [#8]
 
 ## 1.7.1.rc1 [#](https://github.com/enova/landable/compare/v1.7.0...v1.7.1.rc1)